+++ frontend/gui/src/Router.js
@@ -9,7 +9,7 @@ import SignUp from "./components/auth/SignUp";
 import SignOut from "./components/auth/SignOut";
 import Landing from "./components/Landing";
 import SearchPage from "./components/searchPage/container";
-import QueryDetails from "./components/QueryDetails/QueryDetails"
+import QueryDetails from "./components/QueryDetails/container";
 export default () => {
   return (
     <div>
